Added test task to Nukefile.
There are two groups of tests.  Run them both with "nuke test".

* The tests in the MarkdownTests directory are from John Gruber's
Markdown test suite, which you can get here:
http://daringfireball.net/projects/downloads/MarkdownTest_1.0_2007-05-09.tgz

* The tests in the SimpleTests directory are diagnostic tests that
we are using to isolate problems that we find in the larger Markdown
tests.

Currently all 6 of 6 SimpleTests pass, but 17 of 23 MarkdownTests fail.
Many of the failures are over minor issues, but there are still some
significant problems to fix.