Fixed errors in regex replacement. All tests now PASS.
When we replace patterns matched with a regex, we should
do the replacements from right to left.  To make this easier,
I added an eachInReverse: method to NSArray.  It should be
called on the results of [NuRegex findAllInString:] along
with [NSMutableString replaceCharactersInRange: withString:]
to perform the replacement.  With this change, all of
John Gruber's Markdown tests now pass exactly -- this is
based on a character-by-character comparison of NuMarkdown
and Markdown.pl output.

I may have missed a few places where eachInReverse: should
be used, and there's quite a bit of cleanup and polish that
we can do on the code now that we have a significant set
of regression tests that pass.  But for now, here are the
changes that got us golden.